create token pkg
