+++ pkg/token/doc.go
@@ -1,0 +1,2 @@
+// Package token defines tokens that can be used as identity in RBAC authorization.
+package token
